arp sync song position pointer work. Getting closer. Still one ppqn phase issue when changing sync or turning arp mode from off to on with respect to ppqn setting on MIDI START. There's a arp.divider-1 issue that needs to be consistent

+++ monapollo-firmware/monapollo-firmware/arp.h
@@ -33,7 +33,7 @@ struct arp {
 	
 	uint8_t ppqn_counter; 
 	uint8_t divider;
-	uint16_t song_position; //master beat counter
+	uint32_t song_position; //master beat counter
 	
 	uint8_t display; //temporary holder to put a value in to display for testing
